+++ nodejs/fastWorker.js
@@ -188,6 +188,7 @@ class FASTWorker {
         reqid = reqid || 0;
         const defaultConfig = {
             deletedTemplateSets: [],
+            enableIpam: false,
             ipamProviders: [],
             disableDeclarationCache: false
         };
@@ -249,6 +250,11 @@ class FASTWorker {
                         'which is only an issue if something other than FAST (e.g., config sync) is modifying AS3 config.',
                         'Disabling declaration caching will negatively impact FAST performance.'
                     ].join(' ')
+                },
+                enableIpam: {
+                    title: 'Enable IPAM Support (Experimental/Beta)',
+                    description: 'Enable IPAM features in official F5 FAST templates',
+                    type: 'boolean'
                 },
                 ipamProviders: {
                     title: 'IPAM Providers (Experimental/Beta)',
@@ -884,6 +890,60 @@ class FASTWorker {
             .then(() => schema);
     }
 
+    removeIpamProps(tmpl, requestId) {
+        const subTemplates = [
+            ...tmpl._allOf || [],
+            ...tmpl._oneOf || [],
+            ...tmpl._anyOf || []
+        ];
+
+        if (!this._hydrateCache) {
+            this._hydrateCache = {};
+        }
+
+        return Promise.resolve()
+            .then(() => Promise.all(subTemplates.map(x => this.removeIpamProps(x, requestId))))
+            .then(() => {
+                if (this._hydrateCache.__config) {
+                    return Promise.resolve(this._hydrateCache.__config);
+                }
+
+                return this.getConfig(requestId)
+                    .then((config) => {
+                        this._hydrateCache.__config = config;
+                        return Promise.resolve(config);
+                    });
+            })
+            .then((config) => {
+                if (config.enableIpam) {
+                    return Promise.resolve();
+                }
+
+
+                const schema = tmpl._parametersSchema;
+                const props = schema.properties;
+
+                const ipamProps = Object.keys(props).filter(x => x.endsWith('_ipam'));
+
+                ipamProps.forEach((propName) => {
+                    delete props[propName];
+                });
+
+                if (schema.dependencies) {
+                    Object.entries(schema.dependencies).forEach(([key, value]) => {
+                        value = value.filter(x => !x.endsWith('use_ipam'));
+                        if (value.length === 0) {
+                            delete schema.dependencies[key];
+                        } else {
+                            schema.dependencies[key] = value;
+                        }
+                    });
+                }
+
+                return Promise.resolve();
+            });
+    }
+
     convertPoolMembers(reqid, apps) {
         reqid = reqid || 0;
         const convertTemplateNames = [
@@ -1158,6 +1218,14 @@ class FASTWorker {
                     return Promise.resolve()
                         .then(() => this.checkDependencies(tmpl, reqid, true))
                         .then(() => this.hydrateSchema(tmpl, reqid, true))
+                        .then(() => {
+                            // Remove IPAM features in official templates if not enabled
+                            if (tmplid.split('/')[0] !== 'bigip-fast-templates') {
+                                return Promise.resolve();
+                            }
+
+                            return this.removeIpamProps(tmpl, reqid);
+                        })
                         .then(() => {
                             restOperation.setBody(tmpl);
                             this.completeRestOperation(restOperation);